refactor(repository): modularize by concern and add event_type_breakdown
Split the 540-line EventRepository into three concern mixins under
app/db/repositories/: WriteRepository, ReadRepository,
IntelligenceRepository (all inheriting RepositoryBase). app/db/repository.py
is now a thin 43-line aggregator that re-exports EventRepository — all
existing callers require zero changes.

Adds get_source_ip_event_type_breakdown() to IntelligenceRepository and
wires it into get_source_ip(), completing the event_type_breakdown contract
on GET /api/intelligence/ips/{ip}. 8 new integration tests cover the full
breakdown contract including isolation, empty-dict, and list-exclusion cases.

239/239 tests pass. No callers modified.
